using web pack to run on build
