+++ index.html
@@ -6,9 +6,9 @@
 	<title>Seinfeld Quiz</title>
 	<link rel="stylesheet" href="css/normalize.css">
 	<link href='https://fonts.googleapis.com/css?family=Roboto+Slab' rel='stylesheet' type='text/css'>
-	<link rel="stylesheet" href="build/seinfeld-quiz.1.0.0.css">
+	<link rel="stylesheet" href="seinfeld-quiz.1.0.0.css">
 	<script src="http://code.jquery.com/jquery.min.js"></script>
-	<script src="js/app.js"></script>
+	<script src="seinfeld-quiz.1.0.0.js"></script>
 </head>
 <body>
 	<header class="title">
